feature/ product repository test

+++ backend/src/main/java/com/nextivis/nishopping/domain/entity/Product.java
@@ -37,5 +37,5 @@ public class Product {
     private String publisher;
     private String summary;
     private String size;
-    private Float weight;
+    private int weight;
 }
